# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/createTask": {
            "post": {
                "description": "Crea una nueva tarea en el sistema",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Crear una nueva tarea",
                "parameters": [
                    {
                        "description": "Datos de la tarea",
                        "name": "task",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.CreateTaskRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/api.CreateTaskRequest"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/getAllTasks": {
            "get": {
                "description": "Obtiene una lista de todas las tareas",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Obtener todas las tareas",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api.Task"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/getAllUsers": {
            "get": {
                "description": "Obtiene una lista de todos los usuarios",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Obtener todos los usuarios",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api.User"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/getTasksById/{id}": {
            "get": {
                "description": "Obtiene una tarea específica por su ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Obtener una tarea por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de la tarea",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Task"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/tasks/{id}/comments": {
            "post": {
                "description": "Crea un nuevo comentario en una tarea",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comments"
                ],
                "summary": "Crear un nuevo comentario",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de la tarea",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos del comentario",
                        "name": "comment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.Comment"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/api.Comment"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/updateTasks/{id}": {
            "put": {
                "description": "Actualiza los datos de una tarea existente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Actualizar una tarea",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de la tarea",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos actualizados de la tarea",
                        "name": "task",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.Task"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Task"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/users": {
            "post": {
                "description": "Crea un nuevo usuario en el sistema",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Registrar un nuevo usuario",
                "parameters": [
                    {
                        "description": "Datos del usuario",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.RegisterUserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/api.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/users/login": {
            "post": {
                "description": "Autentica a un usuario y devuelve un token JWT",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Iniciar sesión",
                "parameters": [
                    {
                        "description": "Credenciales (email y contraseña)",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.Comment": {
            "type": "object",
            "properties": {
                "content": {
                    "description": "Contenido del comentario",
                    "type": "string"
                },
                "createdAt": {
                    "description": "Fecha de creación",
                    "type": "string"
                },
                "id": {
                    "description": "ID único del comentario",
                    "type": "integer"
                },
                "taskID": {
                    "description": "Tarea asociada",
                    "type": "integer"
                },
                "userID": {
                    "description": "Usuario que hizo el comentario",
                    "type": "integer"
                }
            }
        },
        "api.CreateTaskRequest": {
            "type": "object",
            "required": [
                "created_by",
                "priority",
                "status",
                "title"
            ],
            "properties": {
                "assigned_to": {
                    "description": "Usuario asignado",
                    "type": "integer"
                },
                "created_by": {
                    "description": "Usuario que creó la tarea",
                    "type": "integer"
                },
                "description": {
                    "description": "Descripción detallada",
                    "type": "string"
                },
                "due_date": {
                    "description": "Fecha límite de la tarea",
                    "type": "string"
                },
                "priority": {
                    "description": "Prioridad (low, medium, high)",
                    "type": "string"
                },
                "status": {
                    "description": "Estado (pending, in_progress, completed)",
                    "type": "string"
                },
                "title": {
                    "description": "Título de la tarea",
                    "type": "string"
                }
            }
        },
        "api.RegisterUserRequest": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password"
            ],
            "properties": {
                "email": {
                    "description": "Correo electrónico",
                    "type": "string"
                },
                "name": {
                    "description": "Nombre del usuario",
                    "type": "string"
                },
                "password": {
                    "description": "Contraseña en texto plano",
                    "type": "string"
                }
            }
        },
        "api.Task": {
            "type": "object",
            "properties": {
                "assignedTo": {
                    "description": "Usuario asignado",
                    "type": "integer"
                },
                "comments": {
                    "description": "Comentarios asociados a la tarea",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.Comment"
                    }
                },
                "createdAt": {
                    "description": "Fecha de creación",
                    "type": "string"
                },
                "createdBy": {
                    "description": "Usuario que creó la tarea",
                    "type": "integer"
                },
                "description": {
                    "description": "Descripción detallada",
                    "type": "string"
                },
                "dueDate": {
                    "description": "Fecha límite de la tarea",
                    "type": "string"
                },
                "id": {
                    "description": "ID único de la tarea",
                    "type": "integer"
                },
                "priority": {
                    "description": "Prioridad (low, medium, high)",
                    "type": "string"
                },
                "status": {
                    "description": "Estado (pending, in_progress, completed)",
                    "type": "string"
                },
                "title": {
                    "description": "Título de la tarea",
                    "type": "string"
                },
                "updatedAt": {
                    "description": "Fecha de última actualización",
                    "type": "string"
                }
            }
        },
        "api.User": {
            "type": "object",
            "properties": {
                "assignedTasks": {
                    "description": "Tareas asignadas al usuario",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.Task"
                    }
                },
                "comments": {
                    "description": "Comentarios hechos por el usuario",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.Comment"
                    }
                },
                "createdAt": {
                    "description": "Fecha de creación",
                    "type": "string"
                },
                "email": {
                    "description": "Correo electrónico (único)",
                    "type": "string"
                },
                "id": {
                    "description": "ID único del usuario",
                    "type": "integer"
                },
                "name": {
                    "description": "Nombre del usuario",
                    "type": "string"
                },
                "passwordHash": {
                    "description": "Hash de la contraseña",
                    "type": "string"
                },
                "tasks": {
                    "description": "Tareas creadas por el usuario",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.Task"
                    }
                },
                "updatedAt": {
                    "description": "Fecha de última actualización",
                    "type": "string"
                }
            }
        }
    }
}